+++ CHANGES.rst
@@ -19,6 +19,8 @@
 
 - Fix waveband option in registry.regsearch [#175]
 
+- Fix to regtap.ivoid2service(), few decode()'s, para_format_desc  was moved to utils.  [#177]
+
 
 0.9.3 (2019-05-30)
 ==================